Adding initial testing harness for build automation with py paver
For operations withing a virtual environment, py paver requires
that the virtual environment would provide a bin/activate_this.py.
This file would normally be installed in an environment created with
the virtualenv package.

In order to ensure that a virtualenv environment can be created during
project boostrapping, the ensure_env command is modified in project.py,
such as to create an intermediate virtual environment using venv for
creating the project's primary virtual environment.

Additional changes
- adding a tmpdir option for project.py, setting TMPDIR if provided,
  creating any intemrediate TMPDIR directories with a limited umask
- adding virtualenv, paver to dev_depends (project.ini => setup.py)
- adding a Makefile parameter and project.py cmd option for passing
  options to the virtualenv cmd